프로그램 기능 구현 중간 저장 수조 정보 (컨트롤러, 서비스, 퍼시스턴스) 수조 상태 정보 (컨트롤러, 서비스, 퍼시스턴스)
기능 별 매퍼 xml문서 작성
수조 테이블 매퍼
수조 상태 정보 테이블 매퍼

+++ src/main/java/kr/co/ffm/system/watertankstatus/WatertankStatus.java
@@ -6,8 +6,8 @@ import lombok.Data;
 public class WatertankStatus {
     private int no;
     private String watertankId;
-    private int temperature;
-    private int ph;
-    private int oxygen;
+    private double temperature;
+    private double ph;
+    private double oxygen;
     private String measureTime;
 }
